+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/ToolWindows/App/TestConnectAppProtocolAndOpenClientCommand.cs
@@ -27,7 +27,6 @@ using Google.Solutions.IapDesktop.Core.ClientModel.Traits;
 using Google.Solutions.IapDesktop.Core.ClientModel.Transport;
 using Google.Solutions.IapDesktop.Core.ProjectModel;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.App;
-using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Rdp;
 using Google.Solutions.IapDesktop.Extensions.Session.Settings;
 using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.App;
 using Google.Solutions.Mvvm.Binding.Commands;
@@ -89,7 +88,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
         [Test]
         public void WhenContextOfWrongType_ThenQueryStateReturnsUnavailable()
         {
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(new Mock<IAppProtocolClient>().Object, null),
@@ -112,7 +111,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
             var client = new Mock<IAppProtocolClient>();
             client.SetupGet(c => c.IsAvailable).Returns(false);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, null),
@@ -133,7 +132,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
             var client = new Mock<IAppProtocolClient>();
             client.SetupGet(c => c.Executable).Returns("doesnotexist.exe");
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, null),
@@ -148,7 +147,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
             var client = new Mock<IAppProtocolClient>();
             client.SetupGet(c => c.Executable).Returns(CmdExe);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, null),
@@ -169,7 +168,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
             client.SetupGet(c => c.IsNetworkLevelAuthenticationSupported).Returns(false);
             client.SetupGet(c => c.IsUsernameRequired).Returns(false);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(
@@ -199,7 +198,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
             var settings = InstanceConnectionSettings.CreateNew(SampleLocator);
             settings.AppUsername.StringValue = "user";
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -234,7 +233,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                     out username))
                 .Returns(DialogResult.OK);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -270,7 +269,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                     out username))
                 .Returns(DialogResult.OK);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -306,7 +305,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                     out username))
                 .Returns(DialogResult.Cancel);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -335,7 +334,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
             settings.RdpPassword.ClearTextValue = "password";
             settings.RdpDomain.StringValue = "domain";
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -373,7 +372,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                     out userCredential))
                 .Returns(DialogResult.OK);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -413,7 +412,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                     out userCredential))
                 .Returns(DialogResult.OK);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -451,7 +450,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                     out userCredential))
                 .Returns(DialogResult.Cancel);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 CreateFactory(client.Object, settings),
@@ -497,7 +496,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.App
                 processFactory.Object,
                 settingsService.Object);
 
-            var command = new OpenWithClientCommand(
+            var command = new ConnectAppProtocolAndOpenClientCommand(
                 new Mock<IWin32Window>().Object,
                 new SynchronousJobService(),
                 factory,
